Created homepage hero and added it to homepage
