+++ pages/index.tsx
@@ -4,7 +4,7 @@ import Link from "next/link"
 import Head from "next/head"
 import Navbar from "../components/layout/Navbar"
 import Testimonials from "../components/sections/homepage/Testimonials"
-import { Hero } from "../components/sections/Hero"
+import { Hero } from "../components/sections/homepage/Hero"
 import { TeamCard } from "../components/elements/TeamCard"
 // import avatar from "../public/avatar.png"
 import { SectionWrapper } from "../components/layout/SectionWrapper"
